bash_aliasesの追加

+++ setup.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-cp vimrc ~/.vimrc
+if [ ! -f ~/.bash_aliases ]; then
+  cp bash_aliases ~/.bash_aliases
+  source ~/.bash_aliases
+fi
+if [ ! -f ~/.vimrc ]; then
+  cp vimrc ~/.vimrc
+fi
 cd ~
 mkdir -p .vim/colors
 cd .vim
